backend/requirements.txt
boto3==1.41.5
botocore==1.41.5
python-docx==1.1.0
fastapi==0.122.0
jose==1.0.0
keras==3.12.0
# numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
scikit_learn==1.7.2
starlette==0.50.0
tensorflow==2.17.0
uvicorn==0.38.0
python-multipart==0.0.9
numpy==1.26.4
